Temporarily use the database for tests
Fix some regressions as well.
