+++ server/games/ladderGamesContainer.py
@@ -148,7 +148,7 @@ class Ladder1V1GamesContainer(GamesContainer):
         gameName = str(player1.login + " Vs " + player2.login)
         
         player1.state = PlayerState.HOSTING
-        gameid = self.games_service.createUuid(player1.id)
+        gameid = self.games_service.createUuid()
         player2.state = PlayerState.JOINING
 
         map_pool = self.choose_ladder_map_pool(player1, player2)
